二分

+++ T050. Pow(x, n).cpp
@@ -3,7 +3,7 @@ public:
     double pow(double x, int n){
         if(n == 0)
             return 1.0;
-        int half = pow(x,n / 2);
+        double half = pow(x,n / 2);
         if(n % 2)
             return x * half * half;
         return half * half;
